refactor: updated /components file structure
Moved /PWA and /Navigation to /root/PWA and /root/Navigation. Also the BackButton was extracted from the /root/Navigation directory and moved the /Shared/Navigation directory.
