+++ app/(components)/root/Navigation/NavigationBar.tsx
@@ -1,10 +1,10 @@
 import ColorModeSwitcher from "@/app/(components)/ColorModeSwitcher"
 import * as React from "react"
-import SessionProfile from "@/app/(components)/Navigation/SessionProfile"
+import SessionProfile from "@/app/(components)/root/Navigation/SessionProfile"
 import ClickableCategory, { ClickableCategoryProps, Item } from "@/components/Shared/Menu/ExpandableMenu"
-import MobileNavbarProvider from "@/app/(components)/Navigation/Mobile/MobileNavbarProvider"
-import MobileNavigationBar from "@/app/(components)/Navigation/Mobile/MobileNavigation"
-import HamburgerMenu from "@/app/(components)/Navigation/Mobile/HamburgerMenu"
+import MobileNavbarProvider from "@/app/(components)/root/Navigation/Mobile/MobileNavbarProvider"
+import MobileNavigationBar from "@/app/(components)/root/Navigation/Mobile/MobileNavigation"
+import HamburgerMenu from "@/app/(components)/root/Navigation/Mobile/HamburgerMenu"
 import { Session } from "next-auth"
 import { session } from "next-auth/core/routes"
