+++ app/layout.tsx
@@ -1,12 +1,12 @@
 import '@/public/globals.css'
 import React from "react"
 import { useColorModeValue } from "@/app/(components)/ColorModeHandler"
-import NavigationBar, { NavigationBarProps } from "@/app/(components)/Navigation/NavigationBar"
+import NavigationBar, { NavigationBarProps } from "@/app/(components)/root/Navigation/NavigationBar"
 import AAULogo from "@/public/AAU_Logo.png"
-import AuthProvider from "@/app/(components)/Auth/AuthProvider"
+import AuthProvider from "@/app/(components)/root/Auth/AuthProvider"
 import { getServerSession } from "next-auth"
 import { options } from "@/app/api/auth/[...nextauth]/options"
-import Pwa from "@/components/PWA/PWA"
+import Pwa from "@/components/root/PWA/PWA"
 import { Metadata } from "next"
 
 export const metadata: Metadata = {
